# backend/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Cliente API",
    "version": "1.0.0"
  },
  "paths": {
    "/clientes": {
      "get": {
        "summary": "List all clientes",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "A list of clientes",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Cliente"
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new cliente",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "cliente",
            "description": "Cliente to create",
            "schema": {
              "$ref": "#/definitions/Cliente"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Cliente created",
            "schema": {
              "$ref": "#/definitions/Cliente"
            }
          }
        }
      }
    },
    "/clientes/{id}": {
      "get": {
        "summary": "Get a cliente by ID",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "A single cliente",
            "schema": {
              "$ref": "#/definitions/Cliente"
            }
          },
          "404": {
            "description": "Cliente not found"
          }
        }
      },
      "put": {
        "summary": "Update a cliente by ID",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "cliente",
            "description": "Updated cliente",
            "schema": {
              "$ref": "#/definitions/Cliente"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cliente updated",
            "schema": {
              "$ref": "#/definitions/Cliente"
            }
          },
          "404": {
            "description": "Cliente not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a cliente by ID",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "204": {
            "description": "Cliente deleted"
          },
          "404": {
            "description": "Cliente not found"
          }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Login to get a JWT token",
        "parameters": [
          {
            "in": "body",
            "name": "login",
            "description": "Login credentials",
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              },
              "required": ["username", "password"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "JWT token",
            "schema": {
              "type": "object",
              "properties": {
                "access_token": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    }
  },
  "definitions": {
    "Cliente": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      },
      "required": ["id", "nombre", "email"]
    }
  },
  "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}